Add serverless-swagger plugin
