+++ product-service/package.json
@@ -27,6 +27,7 @@
     "jest": "^28.1.3",
     "json-schema-to-ts": "^1.5.0",
     "serverless": "^3.0.0",
+    "serverless-auto-swagger": "^2.8.2",
     "serverless-esbuild": "^1.23.3",
     "serverless-webpack": "^5.8.0",
     "ts-jest": "^28.0.8",
