+++ product-service/src/types/api-types.d.ts
@@ -1,0 +1,7 @@
+export interface Product {
+  id: string;
+  title: string;
+  description: string;
+  price: number;
+  image: string, 
+}
